% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{create_datetime}
\alias{create_datetime}
\title{create datetime from year month}
\usage{
create_datetime(year, month)
}
\arguments{
\item{year}{numeric}

\item{month}{numeric}
}
\value{
lubridate datetime object
}
\description{
create datetime from year month
}
