% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{get_monthly_summary}
\alias{get_monthly_summary}
\title{Year-Month total expenses, income, profit}
\usage{
get_monthly_summary(transactions)
}
\arguments{
\item{transactions}{dataframe of transactions}
}
\value{
dataframe
}
\description{
Year-Month total expenses, income, profit
}
