% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{get_most_recent_file_date}
\alias{get_most_recent_file_date}
\title{Parse out date in file name, get most recent}
\usage{
get_most_recent_file_date(all_transaction_files)
}
\arguments{
\item{all_transaction_files}{all available transaction files}
}
\value{
formatted most recent date string
}
\description{
Parse out date in file name, get most recent
}
