% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mint_Utils.R
\name{get_transactions_dir}
\alias{get_transactions_dir}
\title{gets root directory for transaction files}
\usage{
get_transactions_dir(overview_df)
}
\arguments{
\item{overview_df}{general financial data, contains directory}
}
\value{
character for directory path
}
\description{
gets root directory for transaction files
}
