% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mint_Utils.R
\name{getCategoryList}
\alias{getCategoryList}
\title{Title}
\usage{
getCategoryList(transactions, startDate, includeOutlier = FALSE)
}
\arguments{
\item{transactions}{todo}

\item{startDate}{todo}

\item{includeOutlier}{todo}
}
\value{

}
\description{
Title
}
