% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mint_Utils.R
\name{getMonthlyAvgSpend}
\alias{getMonthlyAvgSpend}
\title{Title}
\usage{
getMonthlyAvgSpend(
  categoryList,
  returnDf = FALSE,
  removeIncome = TRUE,
  removeLoans = TRUE
)
}
\arguments{
\item{categoryList}{todo}

\item{returnDf}{todo}

\item{removeIncome}{todo}

\item{removeLoans}{todo}
}
\value{

}
\description{
Title
}
