% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mint_Utils.R
\name{getNetIncome}
\alias{getNetIncome}
\title{Title}
\usage{
getNetIncome(categoryList, incomeStartDate)
}
\arguments{
\item{categoryList}{todo}

\item{incomeStartDate}{todo}
}
\value{

}
\description{
Title
}
