% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_analyze.R
\name{mint_get_projections}
\alias{mint_get_projections}
\title{Create financial projections based on past data and manual adjustments}
\usage{
mint_get_projections(
  transactions,
  account_df,
  config_file,
  forecast_date_range,
  historical_start_date
)
}
\arguments{
\item{transactions}{df of transactions from mint by category}

\item{config_file}{user config file}

\item{forecast_date_range}{date range to forecast}

\item{historical_start_date}{start of historical data}
}
\value{

}
\description{
Create financial projections based on past data and manual adjustments
}
