% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mint_Utils.R
\name{monthly_category_sum}
\alias{monthly_category_sum}
\title{get monthly sum of spending by category}
\usage{
monthly_category_sum(transactions, start_date, include_outlier = FALSE)
}
\arguments{
\item{transactions}{todo}

\item{start_date}{date to keep transactions}

\item{include_outlier}{}
}
\value{

}
\description{
get monthly sum of spending by category
}
