% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{read_most_recent_transaction_file}
\alias{read_most_recent_transaction_file}
\title{read_most_recent_transaction_file}
\usage{
read_most_recent_transaction_file(overview_df)
}
\arguments{
\item{overview_df}{general financial data}
}
\value{
most recent set of transactions
}
\description{
read_most_recent_transaction_file
}
